Bug JCR item not found: fixing a technical problem.
When two several threads are switching a same versioned document to a normal one in a same time, the version history of the document can be broken. In a such case, it is no more possible to display the foreign resources linked to a such document in error.

This correction takes into account this case by catching the JCR 2.0 ItemNotFoundException exception as it is already the case with the jackrabbit NoSuchItemStateException one.