Feature #12866
When refactoring the authentication code, a bug was by inadvertence
introduced: when the control is sent back to Silverpeas with an
empty login, an unexpected error was thrown because the login has
to be non-empty when the authentication process is really
performed. Now, the login is checked before any authentication
process.