Feature #12866
Remove wrappers that were written just to handle the closing of
resources allocated when creating a node store for the repository.
A ResourceCloser has been written to handle the closing of these resources
when Silverpeas is in shutdown.

Fix a bug in the versioning handling in Silverpeas. In Apache Jackrabbit
2, the base version in the history of a versioned node is handled differently
than in Apache Jackrabbit Oak. In Oak the base version refers always the
latest version of a versioned node and it is referenced by the node
itself. Hence, it is required to delete this reference either by
deleting the node or by removing the versionable mixin from it before
any purge of the version history.

Rename:
- DocumentRepository#lock to DocumentRepository#checkout
- DocumentRepository#unlock to DocumentRepository#checkin
- DocumentRepository#unlockFromContentDeletion to DocumentRepository#checkinFromContentDeletion

Fix the environment test with SilverTestEnv regarding the use of TestBeanContainer by
ServiceProvider. Latter uses a thread cache for singleton but this thread cache were
cleared only before the execution of a test instead of just once the test instance is
created.