Feature #13608: allowing specifying the localized data of component descriptions into bundle property files.
In this first step, the current possibility to specify the localization data of a component directly into its corresponding XML descriptor is kept.

But, for each component, the priority is now given to its bundle property file if it exists. If a localized data is specified into bundle property file, this data is taken into account. If it does not exist, then the localization into XML descriptor file is looked up. If it does not exists localized data neither into bundle property file neither into XML descriptor one, then a technical error is thrown.

The localized data are manipulated from the classes having their name prefixed by 'Localized' which are situated into the same package of the models implemented for component XML descriptors. Some of the 'Localized' implementations, which was already existing, have been completed to observe the new priorities. Some have been added.