Feature #13706
Remove the request cache service which used a ThreadCache for doing its work

Rename the cache services as being instead the cache accessors.
The clear of the cache isn't anymore proposed by the accessor or by the cache
accessor provider. The caller has to get the cache from an accessor first to
then invoke the clear method.