Feature #13791
Fix a bug with the Silverpeas server URL for each user domain. It is now
correctly set by default from the httpServerBase URL.

Silverpeas uses now StringTemplate 4. With this new version of
StringTemplate, a new way to define a template is now supported: the
template descriptor. A template descriptor is a template definition in a
file that satisfies the following schema:
<template name>(<template arguments>) ::= <template body>
with template arguments being a comma-list of attributes passed to the
template. All the attributes set to a StringTemplate instance have to be
declared here, even thoses not used in the template body. The advantage
of the template descriptor is that it allows to write idented templating
code.
For more information about the schema, please see the StringTemplate 4
documentation or the Javadoc of SilverpeasStringTemplate.
The previous syntax of StringTemplate is supported as default syntax.
The existing template files are taken as a raw template in a file by
StringTemplate 4. There no change otherwise the following one:
idented templating code isn't supported as such; each space, each
tabulation, each line-feed, etc. are taken as such when rendering the
text from the raw template. This require to adapt existing templates in
customer dedicated projects.