Feature #13791
A light improvment in the message building within the Notification API.
From now on, the modification of the message by the Notification API for
HTML rendering is removed as it's not the goal of this API to decide how
the text has to be rendered at the place of the templating; the layout
and the way the message is rendered is the goal of the templates used to
build any messages to send.

In the unit tests on the user notification API, replace the use of the
FieldMocker by NotificationSettingsMocker to mock the SettingsBundle
class instances of the NotificationManagerSettings static class.