Feature #13861
Rename User#geteMail() to User#getEmailAddress()

Add some missing translations of user property names for LDAP
domains.

Add new properties to a LDAP user domain to indicate if the email
address of a user is sensitive and if an additional user property is
sensitive. When such data is rendered, an icon is displayed alongside
the property value to indicate the data is a sensitive information.