Feature #13861
The support of sensitive information is only for LDAP domains.
It is based upon two mechanismes:
- the first one is to define the user information that could be
  sensitive. This is done directly into the user domain descriptor.
- the second one is to specify in a per user basis these information are
  actually sensitive, id est information become confidential for the
  user.
Hence, an information defined as potentially sensitive isn't by default
privacy. The administrator has to specify explicitly these information
are sensitive for a user.

In this first version of the feature, there is no mechanism to enable
or disable sensitive information support for a LDAP domain. The
mechanism is on and the email address is automatically considered as a
a potential sensitive information. Even whether the email address is
used as user login and that login follows a name/surname entreprise
convention.

Each time a user profile is updated, its index is also rewriten. The
indexation is used to be able to search a user in Silverpeas. So, when
the data sensitivity is enabled for a user, when reindexing, its email
address and all of the domain properties marked as sensible won't be
indexed. So, the user won't be fetched by querying on one of his
sensitive information.

Refine and fix according to sonarlint the arraypane items.
Take into account sonarlint feedbacks on JobDomainPeas type objects and
on the Administration and AdminController classes.
Take into account the Idea IntelliJ feedback on some classes in user
notification and in adminitrative tasks.

Add additional attribute to the array cell checkbox:
- readonly to disable the user interaction with the checkbox
- action to pass javascript callback on events on the checkbox

Add in the user list of a LDAP domain the information about the
enabling of the data sensitivity for each of them. A sensitive
information is an information that must not be disclosed. It is a
privacy data (a synomym to sensitive).

New function in the menu of the user content page: enable/disable
the sensitive data of a user. This function is taken into account by the
administration service behind. When the user has sensitive data, a
message informing about this is displayed.

Add a new function in the menu of the domain to display the users with
sensitive data. Within this new page, administrators can disable the
sensitivity of the data per user.

As all the users information is got through the OrganizationController,
add a new centralized method in its implementation class to remove any
sensitive data from the user information in the case the requester
isn't an administrator or a manager of the domain to which the
requested user belong. By doing this, any call to a web service or to a
controller or to a service to get information about one or more users
will be treated accordingly.
The OrganizationController implementation delegates calls to the
Administration object. The dependency should be in the direction of
OrganizationController -> Administration. This is why, any reverse
dependencies (id est from Administration to OrganizationController and
mainly those indirect) are removed and replaced by a direct invocation
of a method of the Administration object itself.
As the user details fetched by the Administration single object are
cached, only copies of them are returned by the the single object. This
is for ensuring the sensitive information are only disclosed for the
users with the correct role.