Feature #13961
The goal here is to remove the code, duplicate with Kernel lib, and
applies the modification in Silverpeas Core using this code so that
all the code in Silverpeas Core uses the one from Silverpeas Kernel.
Some classes in Silverpeas Kernel differ to their counterpart in
Silverpeas Core. The implementation of the code in Silverpeas Kernel
replaces those in Silverpeas Core API. So, some unit and integration
tests are modified to work with the new implementation of the codes
from Silverpeas Kernel.

For doing, a first step was to move the duplicate code with Kernel to
a package org.silverpeas.kernel, with Idea IntelliJ, so that all the
dependencies are also modified accordingly. Then, the duplicated code
is removed from Silverpeas Core, and Silverpeas Kernel is declared as
a dependency. Finally, all is checked with the unit and integration
tests and by testing by hand with a running Silverpeas.

What is removed and replaced:
- all the annotations common with those in
  org.silverpeas.kernel.annotation
- all the exceptions common with those in
  org.silverpeas.kernel.exception and in org.silverpeas.kernel
- all the utility classes common with those in
  org.silverpeas.kernel.util
- the cache API common with org.silverpeas.kernel.cache
- the BeanContainer (but the one in core-api differs from the one in
  kernel
- the Logging API that is defined in org.silverpeas.kernel.logging. The
  logging API in Silverpeas Kernel uses SilverpeasResourceLocation to
  get the location of the root directory of the loggers configuration
  files.
- the Bundle API in org.silverpeas.kernel.bundle. This API allows to
  load both configuration properties and l10n bundles through
  ResourceLocator. ResourceLocator uses SilverpeasResourceLocation to
  get the root directory of the configuration properties and the root
  directory of the l10n bundles.

An implementation to the SilverpeasResourceLocation is added and
available for Java SPI: ResourcesRootDirectoriesLocator.

Fix a bug with the configuration of some loggers: namespace duplication.

Take into account the change in unit test with the EnableSilverEnvTest
JUnit 5 extension in Silverpeas Kernel.
Move unit tests from Junit 4 to Junit 5. Removes some Junit 4 rules.
Refactor some unit tests to avoid the setting of static fields of
business objects (DisplayI18NHelper for example)

Remove the ServiceProvider#getSingleton methods as the caching of
singletons is now automatically performed by ManagedBeanProvider. This
avoid any error of caching not singleton beans (when the scope of their
class has changed).

Depends now on Silverpeas Kernel version 1.0.