Feature #14007
Now, in the descriptor of the LDAP domains (and in their corresponding
authentication descriptor), a new property is defined to indicate
whether the credentials used to access the remote directory service
are encrypted: database.encryptedCredentials (and
authServer?.encryptedCredentials (with ? as being the server index)).

In the case the credentials are encrypted, the LdapConfiguration
decrypts both the user DNS and password on demand. For doing, it uses
the Content Encryption API.

In order to make it work, the plateform administrator has to generate a
256 bytes key in hexadecimal. With this key, it has to encrypt explicitly
and by hand both the user DNS and password and write them in the
corresponding properties in the domain and authentication descriptors of
the remote directory service. Then, the 256 bytes key has to be set in
the backoffice of Silverpeas (BackOffice > Tools > Content Encryption)