Feature #14007
Update some code according to sonar analysis feedback.

Add an iterator on the LDAP domains for the encryption of their directory
access credentials. This iterator is registered in the Silverpeas
Encryption API to be invoked when a cipher key is added or updated. For
doing, the Silverpeas Encryption API has been a little updated to take
into account the iterators when a key is added the first time.

The iterator is registered by an initialization service at Silverpeas
bootstrapping. The iterator uses a new object, DomainDescriptor, to
access both the domain definiton and the domain authentication
properties and to update them (to replace the credentials value by their
(new) encrypted counterparts.

So, now, the plateform administrator has just to perform three actions
to encrypt the credentials of a LDAP domain:
* he updates the properties database.encryptedCredentials and
  autServer?.encryptedCredentials to indicate the credentials for
  accessing the directory are encrypted.
* to a running Silverpeas, he registers his cipher key (a 256 bits AES
  key) => Silverpeas updates automatically the LDAP domain properties
  files.
* then, to avoid possible connexion errors for users of the LDAP domain,
  he restarts Silverpeas in order the encrypted credentials are taken
  into account.