Feature #14033
The method Administration#getGroupByNameInDomain doesn't take into
account the specified domain is the mixed one. This peculiar domain
hasn't driver so specific code need to be written to take into account
of it. In the case the asked group is one of the mixed domain, the
method now search directly in Silverpeas without passing through a
domain driver to get the specific identifier of asked group in the
domain.

ChatGroupUserLinkEventListener listens for user added into a group to
register this user in the remote chat service in the case this group is
one of those for which chat service is allowed. Nevertheless, it doesn't
listen for user removed from an allowed group to unregister it from the
remote chat service. Later is now accomplished.

Add a new system event type: recovery. This kind of events is for
removed resources that are recovered from the bin. Now the listener of
system events provides a method to take into account such type of event.

Add system event notification for users group removing and recovery.