Feature #14023: synchronize synchronized group on new user when 'new user registration' feature is enabled
Adding a new parameter `registrationSynchroGroupEnabled` into `authenticationSettings.properties` file in order to disabled if needed the synchronization in a such context.

The processing of synchronizing all synchronized group has been extracted from scheduler in charge to trigger the synchronization. Now, it exists a manager of synchronized group which is invoked by the scheduler and also by the new user registration services.