Bug #14100
Fix bug: when no non synchronized calendars are visible while asking to
create a new event, by default, then, all non synchronized calendars are
proposed for the new event. Otherwise, the usual behaviour is played:
all the visible non synchronized calendars are proposed (or the default
one if one of them is the default calendar).