Bug #14158
In the Drag&Drop Web Edition API, the images as components (a GrapeJS
components as GrapJS is used here as the front framework for the API)
are treated like attachments to the contribution for which the API is
used as editor. But, in the case of a thumbnail uploaded by the user,
the image is just a simply link that can be broken when the original
thumbnail is deleted or replaced by another image. In order to avoid
such issues, and to conform to the way images are treated in the API,
now the uploaded thumbnails are now transformed as a simple image
component, id est as an attachment to the currently edited
contribution through the Drag&Drop Web Edition API.