Fix bug #14219
The input stream on the body of the Http request is buffered only in the
case of a web entity is carried within it (JSON or XML entity). For any
others data type, no bufferization if done.
Beside that, when buffered, ensure the all the attributes from the
decorated HttpRequest are correctly copied.