Bug #14254
In the backoffice, the VueJS component driving the management view of
the spaces and components expects the current app instance id to be set
when the view body is on the rendering of a given app instance.

But, in the case of a creation, update or rank move of an app instance,
once the operation done, when the control flow is passed to render the
app instance info page, the current app instance id isn't set and then
the VueJS component redirect the rendering of the view body to the
parent space info page instead.

In order to render the app instance info page in the context of these
operations, it is required to redirect to a switcher: a web page in
which the control flow is passsed to a JS component controlling the
backoffice view (spAdminWindow) that will then update indirectly the
VueJS component and hence the view body and space/app navigation tree.