Bug #14256
The SilverpeasBeanDAO interface allows to pass directy an SQL where
clause to filter the beans to get or to remove. This is an open door to
SQL injection and we cannot account to the UI or to the service layer to
check the where clause. Some of the where clauses come from directly the
web layer and as such a user can rewrite the clause before this one is
sent to Silverpeas.

In order to avoid this vulnerability, the direct where clause is
replaced by a BeanCriterion that imposes the business layers to decode
the expectations coming from the web layer into a BeanCriterion before
passing it to the SilverpeasBeanDAO object.

Because, the value of a criterion can be subquery dedicated to attempt
an attack, BeanCriterion uses a PreparedStatement in which the
parameters are set by using the corresponding interface of the
PreparedStatement. This would prevent such attempts because it encodes
the textual values.