Bug #14148
The search by a form is performed against the indexes. When a field of a
form is of type PdC, the field is valued with one or more positions on
the PdC by using two differents separators (one to separate two values
of a given position and another one to separate two different
positions) and this value is then indexed/searched under the index
attribute '<name of the form>##Positioning'. Hence, the search on a such
field is done by looking for indexes with such an attribute having a
value matching exactly the one searched. This means the index entries
in which the attribute contains at least the searched value aren't
returned by the Lucene search engine.

In order to fix this, for the PdC field, the value to search is
rewritten so that to be a regexp understandable by Lucene. In order to
centralize this rewritting, all the code building a QueryDescription for
a search by a form (coming from an XML publication template) is now
centralized into the QueryParameters object. Hence this object is used
both by the search on the PdC (by a search form of a publication
template), and by the Users Directory component.

Fix a bug in the search on WYSYWYG contents of a form field. The bug
comes from a mismatch between the name of the component managing the
users in Silverpeas (and used by the indexation) and the name of the
Directory component when using the publication templates to define
additional information about the users. Now, the name of both is "users"
and this name is used both by the user indexation and by the user
directory. For doing, a migration step for the FormTemplate component
has been defined.

Take the occasion also to improve some codes (as reported by
SonarCloud).