Bug #14387
In anonymous mode, the redirection to the contribution referred by a
notification wasn't done correctly when the user isn't authenticated:
* if the user isn't connected, it recieves an access forbidden (normal
  as it is not yet connected)
* and once he signs in, he's redirected into the main page instead of
  the contribution.

Now, when the user accesses from the notification (email) to Silverpeas,
he's automatically redirected into the login page for authentication.
Once authenticated, he's redirected into the page in which is displayed
the contribution referred by the notification.
For doing, in the authentication process, in the session is opening,
the computation of a possible redirection URL takes into account now of
the parameters of redirection. Those parameters are set by
AutoRedirectServlet when the user accesses first, from the notification
(email), to Silverpeas. In our case, this is this servlet that redirects
the user to the login page.

In anonymous mode, if the contribution targeted by a notification is
protected, the user is first automatically redirected to the login page
for authentication, as described above. Some times the protection of the
contribution comes from its parent container (for example, a topic with
specific rights for a publication in a Kmelia): take into account these
peculiarities.