Bug #14397
In JobStartPageRequestRouter, when creating a root space, no space has
been yet set as current space (because there is no parent space to a
root space): this causes a NullPointerException. It was a regression
introduced by the fix of the bug #13761. Restores the check of nullity
of the current space in the setSpaceInfosInRequest method.