Bug #14471
A publication is a peculiar contribution in Silverpeas as it can be
located in several locations in the same or in different component
instances. In this case, we have the original publication along with its
all aliases. This capability is used in the Kmelia application. As
consequence, the publication is only referred uniquely by its local
identifier (its local identifier is its global one), whereas the others
types of contributions are identified uniquely by their global
identifier that is made up of their local identifier and of the
identifier of the application instance they belong to.

Despite this peculiarity, the indexation, the search and the
access rights computation treat the publications only by their local
identifier, and this can cause troubles when the aliases of a
publication (with the original one) are implied. For example, if an
alias was removed, then the index in which the alias is referred is
simply deleted, despite the fact there is others aliases or, worse,
there is the original publication referred by the index.

In the fix, we ensure now that all the treatments are done on both the
local identifier of the publication and the identifier of the
application instance it belongs to. And, in the management of the
indexes of the publications, they are deleted only if there is no
more publications (original one/aliases) referred by them.
And, in the access rights computation, in the case of a search,
we look for the access rights of nodes with specific rights in which
is linked the in-checking alias.