Bug #14503
The move of a node was done by deleting and then recreating in database
the node with its new location, and this without updating its
translations to refer the correct new node.

Replace the deletion and then the creation of the moved node by a simple
suitable update: in fact, a move method already existed in the NodeDAO
but it wasn't anymore used. This method update accordingly a node
details in the database for a node move. The translations of the node
are then not lost.

Refine the update of a node in DefaultNodeService by avoiding to
retrieve again from the database the current node before update; Get it
at first, then use it all along the update process.