Bug #13273
Rename SpaceInst#getSpaceProfileInst to
SpaceProfilInst#getDirectSpaceProfileInst because this method returns
only the direct space profile instance with the specified name and not
all the profile instances (both inherited and not inherited).

Add new method to get both the inherited and direct space profiles with
a specified name: SpaceInst#getSpaceProfileInst(String)

In DefaultAdministration:
- when a new space is created, in the case of a inherited rights
  access, the inherited space manager profile is now taken into account
  in same way than the others inherited space profiles.
- when getting the space profile of a given space and with the specified
  role name, take into account the inherited space manager profile has
  been previously cached

In SpaceResource, when getting the space profiles of a given space with
requesting role names, invokes, through the OrganizationController, for
each role, the Administration#getSpaceProfile(String, SilverpeasRole)
method.