Feature #14464
In calendars, when a notification is sent to users about an operation
done on an event, if this event spans over several days, its end date is
specified in the notification text.

In order to know if an event spans over several days, a new method is
added to the Period object and this method gives this information.