With Silverpeas-Kernel, the parameters passed in LocalizationBundle#getStringWithParams are now encoded for HTML when they are String objects. Take into account of this new capabilities in Silverpeas. This means the HTML specific characters in the parameters (like &, > or < for examples) will be encoded in their entity number counterpart (&amp;, &gt; or &lt; in our examples); this shouldn't cause any issue because the destination of such localized messages are dedicated to be rendered in HTML (even for emails).
In Web messages, in order to avoid to encode for HTML all the message
text, improve the MessageNotifier#format private static method to apply
the encoding for HTML only on String parameters as those are usually
passed by the user himself when authoring or editing a contribution.

In user notifications, encode any possible text passed by a user in
Silverpeas (the title and the description of a contribution for
example). Take care of the difference between the direct notification
about a contribution by a user and the user notification mechanism when
a contribution is modified or created.