Feature #3990
A group of users can now be copied/cut and then pasted either as a root
group of the domain of users it belongs to or as a subgroup of another
group of users in the same domain of users. The cut&paste of a group
result to a move of this group. A group cannot be paste to where there
is another group with the same name (groups at the same level have to be
unique by their name).
Control has be added to allow only admins to be able to copy and to move
groups in a given domain of users.
Only the Silverpeas domain, the mixed domain, and the SQL-backed domains
support the copy/move of group of users.