Feature #3990
In the clipboard.jsp web page template, replace the programmatic
construction of the array pane by its declarative approach with the
corresponding taglibs. For doing SilverpeasKeyData has been modified in
order to have all the properties required for the array pane content
rendering; the SilverpeasKeyData object is then used instead of the
IndexEntry object (SilverpeasKeyData is for the clipboard's item whereas
IndexEntry is for the indexation).