Bug #14491
In uploadFile.jsp, encode for HTML the browsbar info in the breadcrumbs.

In the tree menu building, the tree filter implementations are now
retrieved by CDI and not anymore through the classloader.

Fix the code of the tree menu building to please Sonar.

Fix the bug: the sorting of the subspaces is asked.