Feature #14464
Rename the interface Plannable to Planned as it is about yet planned
temporal objects.
Add two new interfaces:
- PlannedOnCalendar to define Planned temporal objects that are planned
  on a given calendar;
- PlannableOnCalendar to define PlannedOnCalendar objects with the
  additional capability to be planned on a given calendar.
CalendarEvent is a PlannableOnCalendar whereas CalendarEventOccurrence
is a PlannedOnCalendar: a calendar event, when created, has to be
planned on a calendar in order to be serialized and a calendar event
occurrence is an instance of a given calendar event occurring at some
time in a calendar (it is then yet planned on the calendar).

Because CalendarEvent and CalendarEventOccurrence implements the
PlannedOnCalendar interface, it is more easy to handle both without
knowing their concrete type. And then, the methods in which both are
used without any discernation are redefined to handle PlannedOnCalendar
objects instead of just more general Contribution ones.

For the user notification about a calendar event, fix the computation
of the end date of the event/event occurrence in the case the period
of such an event/event occurrence is in days (and not in days and time).