Fearture #14738
Change the signature of the callback in CDIResourceEventListener.
Take into account the change in the signature of the methods in
CDIResourceEventListener in the implementors.

Add a mecanism to notify about the removal of a user in some roles he
played (either explicitly or through a user group) in some component
instances. For doing, two listeners are added:
- ProfileInstUpdateEventListener to listen for update in the list of
  users and groups of a given access right profile of a component
  instance.
- GroupUserUnlinkEventListener to listen for removal of a user from a
  user group (this group could play some roles in some component
  instances).

Add integration tests covering this new mechanism.