Bug #13477
Fix the bug: now the publications aliases cannot be shared to other
people for security reasons.

Indeed, the publication aliasing covers two features:
* A way to allow access for publications to users having no access
  for the EDM in which those publications are located.
* A way to multi categorize a publication;
For the first point, for security reasons, the users have only read
rights on the publication aliases. They have then all the functions
mapped to their role that don't break this read-only constraint. Only
users having access for the EDM or the folder in the EDM in which the
alias has been created can then access, through the alias, for the
publication.
For the second point, the multi categorization feature is limited by
the way the publications cannot be directly modified in those different
categorizations; they cannot be modified by their aliases. In fact, the
feature is more a consequence of the first one than a thinkfully one.

Beside this, the folder or publication sharing is a way to share
information both to other users in Silverpeas and external users. This
is why the feature can be a way to counterpass the access constraint on
publications with aliases. So, this fix ensures now the publication
aliases cannot be shared to other people than those expected by the
contributor who created the aliases.