Feature #10785
Implement the feature by following the same mechanism used for both
removed users and groups: the delay in days since the removal of a space
or of a component instance after which this space or application is
definitively deleted is specified into the
org/silverpeas/admin/admin.properties configurartion file. Two delays
and two CRONs can be set: one for the removed spaces and another one for
the removed applications.