Feature #10786
Rename the methods PublicationService#removePublication and
NodeService#removeNode to respectively
PublicationService#deletePublication and NodeService#deleteNode as these
methodds delete definitively the contribution. This is to introduce a
new state to these contribution types: the removal one.

Add a new business function for the publications and the nodes: the
removal. The function marks the publication or the node as being
removed; it is an intermediate state of the contribution before their
definitive deletion. This new state is dedicated to the applications
like Kmelia in which these contributions can be placed into a bin in
order to be either restored or deleted.

For doing add two new attributes in PublicationDetail and in NodeDetail
classes: removalDate and removerId for respectively the date of the
contribution removal and the unique identifier of the user invoking
this removal. Corresponding to these attributes, add the SQL fields.
Latter will drive to a new database migration step for the publications
and for the nodes.