Bug #14829
Refactor the credentials management API in order the handlers can
indicate if credentials pre-checking treatment has to be done before
invoking them. The refactor uses the Visitor pattern.

Now, when the user request a password reseting or a password change, if
the pair login/domain doesn't exist in Silverpeas, an error page is
displayed. The HTT response to the navigator doesn't change if there is
an error or not; only the content of the response (the document) change.
The error page instead of talking about the not found login identifier
for a given user domain, informs just the user its password change/reset
isn't allowed.