Bug #14829
Now it isn't anymore possible to change his password through the login
form. This feature has been removed for security reasons.
To change his password, the user has either to reset it in the login form
(if this feature is enabled) or to change it in his profile page once
signed in Silverpeas.

When reseting his password with an invalid login id, the same message is
given than with a valid login id. So nobody cannot know if a user with
such a login id exists or not.