Feature #10786
Take into account the new properties, removalDate and removerId, in
NodeDetails and in PublicationDetails, in the persistence of these
objects.

Refactor PublicationDAO to be managed bean instead of a static class.

WAComponentRegistry is now the more priority initialization service; it
will be then executed (to load all the available Silverpeas apps) among
the first important initialization services.