Bug #14787
Add a new sys event notifier for changes in the user roles for a given
component instance. The related event is UserRoleEvent.
Two listeners are added for this:
- one for events about a user being removed from a user group
- one for events about a change in the right profile of a component instance
The first one figures out all the roles the group plays in all the
component instances in order to send a notification about the user not
playing anymore the role played by the group.
The second one figures out the users who are removed from the right
profile either directly or through a group removed from the right
profile in order to send a notification about those users not playing
anymore the role related by the right profile.

Add a listener of UserRoleEvents in order to figure out, for each
concerned component instances, and for their nodes having local right
profiles, the users related by the event and who play a role for the
nodes. For those users, and in the case they don't play another role in
the component instance, they are removed from the right profiles of the
nodes.

Add a listener of ProfileInstEvent in order to spread to the nodes with
local right profiles the remove of any user groups.

The two listeners above invoke the NodeProfileInstUpdater service to
synchronize the changes of user roles in the component instance (and
then also in the space of the component instance) to the nodes with
specific local access rights.