Bug #14787
Report the fix from branch 6.4.x:
Two new listeners are added for the Node API. The first one listens for
events about change on the roles played by a user in a component
instance (either directly or through a group playing a role) and carried
by the UserRoleEvent event. The second one listens for events about
change on the roles played by a group in a component instance and
carried by the ProfileInstEvent event.
The goal of the two listeners is to figuring out the users and the
groups that don't play anymore a role in a component instance and then
invoke the NodeProfileInstUpdater service to remove from the specific
access rights of nodes the related users and groups.