Bug #14869
NodeService#createNode methods return now the new NodeDetails instead of
the NodePK of the created node.

Check directly the nullity of the returned ProfileInst object after the
invocation of ProfileManager#getProfileInst(...) method in order to
avoid to spread the nullity possibility of the ProfileInst object up to
the chain of call.