Bug #14807
In the commenting mechanism, when the content of the message and the
additional message are the same, the additional message is dismissed in
the sent notification. But, in the case of the notification about a
comment, the additional message is the comment as is whereas the content
of the message is the comment escaped for its printing in HTML! So, the
two becomes different.
In the case of a comment, there is no need to have both the additional
message and the content of the notification message as they are
identical. The fix here is to remove the extra message setting in the
case of a notification about a comment.